+++ CHANGELOG.md
@@ -404,6 +404,7 @@ Notable features:
 - Fix segmentation fault in E2E when light-client feature flag is enabled. [PR](https://github.com/prysmaticlabs/prysm/pull/14699)
 - Fix `searchForPeers` infinite loop in small networks.
 - Fix slashing pool behavior to enforce MaxAttesterSlashings limit in Electra version.
+- Fix the validation issue before broadcasting a voluntary exit message: use wall clock validation.
 
 ### Security